The vector embeddings calculated for each line item is now normalized. Changed vectorstore policy in config.py from cosine to dotproduct.
The normalization change is welcome in case we use the vectors for any operation using dotproduct (possibly during a future re-ranking step).
We can also switch from azure vectorstore policy cosine to dotproduct. (We want to use cosine, as we are not interested in vector magnitude but dotproduct is the same as cosine as long as the vector is normalized. I have a feeling this will be less computationally expensive, but this depends on the Azure-side implementation.
Usually if I would implement cosine vs dotproduct vector comparisons, cosines would be more expensive computationally because it involved trignometry, while in dotproduct you just mulitply and add the terms.)